+++ example/lib/todo/todo_cube.dart
@@ -2,7 +2,7 @@ import 'package:cubes/cubes.dart';
 import 'package:examplecube/todo/todo_register.dart';
 
 class TodoCube extends Cube with CubeNavigation {
-  final todoList = ['first TODO', 'second TODO'].obsValue;
+  final todoList = ['first TODO', 'second TODO'].obs;
 
   void addTodo(String todo) => todoList.add(todo);