+++ lib/src/injector/injector.dart
@@ -11,24 +11,19 @@ typedef CDependencyInjectorAsyncBuilder<T> = Future<T> Function(
 
 /// Interface responsible to manager dependency injector
 abstract class CInjector {
-  /// Method used to register dependency
-  void putDependency<T extends Object>(
-    CDependencyInjectorBuilder<T> builder, {
-    String? dependencyName,
-    DependencyRegisterType type = DependencyRegisterType.factory,
-  });
-
   /// Method used to register dependency by lazy
   void putLazySingleton<T extends Object>(
     CDependencyInjectorBuilder<T> builder, {
     String? dependencyName,
   });
 
+  /// Method used to register dependency by factory
   void putFactory<T extends Object>(
     CDependencyInjectorBuilder<T> builder, {
     String? dependencyName,
   });
 
+  /// Method used to register singleton dependency
   void putSingleton<T extends Object>(
     T value, {
     String? dependencyName,