+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [next]
+- Renamed `obsValue` to `obs`
+- Improvements in injector, now you have separated methods to register your dependency:
+    - `putSingleton`
+    - `putLazySingleton`
+    - `putFactory`
+
 # [1.4.0]
 - little improvements
 - renamed `Cubes.registerDependency` to `Cubes.putDependency`;